# go/docs/openapi/swagger.yaml
basePath: /
definitions:
...
  entity.Playlist:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_queue:
        type: boolean
      owner_id:
        type: string
      title:
        type: string
      track_count:
        description: maintenu par trigger BDD (migration 005)
        type: integer
      tracks:
        items:
          $ref: '#/definitions/entity.Track'
        type: array
    type: object
...
  entity.Track:
    properties:
      added_at:
        type: string
      artist:
        type: string
      created_at:
        type: string
      duration:
        description: secondes
        type: integer
      file_url:
        description: URL Supabase Storage
        type: string
      id:
        type: string
      playlist_id:
        description: Champs de liaison playlist (remplis quand la track est dans une
          playlist)
        type: string
      position:
        type: integer
      title:
        type: string
      uploaded_by:
        type: string
    type: object
...
  handler.AddTrackRequest:
    properties:
      track_id:
        type: string
    required:
    - track_id
    type: object
...
  handler.CreatePlaylistRequest:
    properties:
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  handler.ReorderRequest:
    properties:
      track_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - track_ids
    type: object
...
  handler.UpdatePlaylistRequest:
    properties:
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    name: StreamPulse team
    url: https://github.com/Ghost-15/streaming
  description: |-
    Live audio streaming API. A broadcaster ingests an audio
    stream, the Hub fans it out to every listener, and playlists,
    favorites and listening history are exposed as REST resources.
  license:
    name: Educational project
  termsOfService: https://github.com/Ghost-15/streaming
  title: StreamPulse API
  version: "1.0"
paths:
...
  /api/v1/playlists:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Playlist'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List playlists owned by the authenticated user
      tags:
      - playlists
    post:
      consumes:
      - application/json
      parameters:
      - description: Playlist payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePlaylistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Playlist'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new playlist
      tags:
      - playlists
  /api/v1/playlists/{id}:
    delete:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a playlist
      tags:
      - playlists
    get:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Playlist'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a playlist by ID
      tags:
      - playlists
    put:
      consumes:
      - application/json
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Playlist'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a playlist title
      tags:
      - playlists
  /api/v1/playlists/{id}/next:
    post:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Track'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pop the next track of a playlist queue
      tags:
      - playlists
  /api/v1/playlists/{id}/tracks:
    post:
      consumes:
      - application/json
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Track payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddTrackRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a track to a playlist
      tags:
      - playlists
  /api/v1/playlists/{id}/tracks/{trackID}:
    delete:
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Track ID
        in: path
        name: trackID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a track from a playlist
      tags:
      - playlists
  /api/v1/playlists/{id}/tracks/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Ordered track IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ReorderRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reorder the queue of a playlist
      tags:
      - playlists
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Access token issued by /api/v1/auth/login, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"